+++ lib/segment/screen.ex
@@ -1,0 +1,15 @@
+defmodule Segment.Screen do
+  @derive [Poison.Encoder]
+  @method "screen"
+
+  defstruct [
+    :userId,
+    :name,
+    :properties,
+    :context,
+    :timestamp,
+    :integrations,
+    :anonymousId,
+    method: @method
+  ]
+end
